+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of the jenkins cookbook.
 
 ## Unreleased
 
+- Move credentials_file to a custom resource
+- Add the credentials partial for reuse for all credentials resources
+
 ## 9.5.2 - *2022-03-28*
 
 - Fix permissions on reusable workflow